Revert "vo_opengl: make the size of interpolation textures exact"
This reverts commit 0bf87d7b2199c56e4f1b74a6bd2184c470a3bb56.

Reallocating the FBOs on every resize is very slow. It affects resizing
the window, as well as changing the video size itself with e.g.
panscan. Since the original change was done based on a single user
complaint, but the change itself caused a lot of complaints, we decided
to just revert it.